pyproject.toml
[tool.poetry]
name = "raytorch"
version = "0.1.0"
description = "Ray tracing using PyTorch"
authors = ["Boston Walker"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^9.0.0"
torch = "^1.10.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
